+++ daengle-groomer-api/src/main/java/ddog/groomer/application/AccountService.java
@@ -38,9 +38,8 @@ public class AccountService {
 
     @Transactional
     public SignUpResp signUp(SignUpReq request, HttpServletResponse response) {
-        if (this.hasInvalidGroomerDataFormat(request)) {
+        if (this.hasInvalidSignUpReqDataFormat(request))
             throw new GroomerException(GroomerExceptionType.INVALID_REQUEST_DATA_FORMAT);
-        }
 
         Account newAccount = Account.create(request.getEmail(), Role.GROOMER);
         Account savedAccount = accountPersist.save(newAccount);
@@ -74,12 +73,15 @@ public class AccountService {
 
     @Transactional
     public void modifyInfo(ModifyInfoReq request, Long accountId) {
+        if(this.hasInvalidModifyInfoReqDataFormat(request))
+            throw new GroomerException(GroomerExceptionType.INVALID_REQUEST_DATA_FORMAT);
+
         Groomer groomer = groomerPersist.getGroomerByAccountId(accountId);
         Groomer updatedGroomer = GroomerMapper.withUpdate(groomer, request);
         groomerPersist.save(updatedGroomer);
     }
 
-    private boolean hasInvalidGroomerDataFormat(SignUpReq request) {
+    private boolean hasInvalidSignUpReqDataFormat(SignUpReq request) {
         try {
             Groomer.validateShopName(request.getShopName());
             Groomer.validateName(request.getName());
@@ -92,6 +94,16 @@ public class AccountService {
             return false; // 모든 유효성 검사 통과
         } catch (IllegalArgumentException e) {
             return true; // 유효성 검사 실패
+        }
+    }
+
+    private boolean hasInvalidModifyInfoReqDataFormat(ModifyInfoReq request) {
+        try {
+            Groomer.validateIntroduction(request.getIntroduction());
+
+            return false;
+        } catch (IllegalArgumentException e) {
+            return true;
         }
     }
 }
