+++ daengle-groomer-api/src/main/java/ddog/groomer/application/exception/AccountException.java
@@ -1,6 +1,7 @@
 package ddog.groomer.application.exception;
 
-import ddog.groomer.application.exception.common.CustomRuntimeException;
+
+import ddog.auth.exception.common.CustomRuntimeException;
 
 public class AccountException extends CustomRuntimeException {
     public AccountException(AccountExceptionType type, Object... args) {
