+++ daengle-groomer-api/src/main/java/ddog/groomer/application/exception/GroomerException.java
@@ -1,6 +1,7 @@
 package ddog.groomer.application.exception;
 
-import ddog.groomer.application.exception.common.CustomRuntimeException;
+
+import ddog.auth.exception.common.CustomRuntimeException;
 
 public class GroomerException extends CustomRuntimeException {
     public GroomerException(GroomerExceptionType type, Object... args) {
