--- daengle-groomer-api/src/main/java/ddog/groomer/application/exception/common/IllegalParameterException.java
@@ -1,7 +1,0 @@
-package ddog.groomer.application.exception.common;
-
-public class IllegalParameterException extends RuntimeException {
-    public IllegalParameterException(String message, Object... args) {
-        super();
-    }
-}
