--- daengle-groomer-api/src/main/java/ddog/groomer/application/exception/common/NotFoundException.java
@@ -1,9 +1,0 @@
-package ddog.groomer.application.exception.common;
-
-import org.springframework.http.HttpStatus;
-
-public class NotFoundException extends CustomRuntimeException {
-    public NotFoundException(String message, Object... args) {
-        super(message, HttpStatus.NOT_FOUND, null);
-    }
-}
