+++ daengle-payment-api/src/main/java/ddog/payment/application/exception/OrderException.java
@@ -1,7 +1,7 @@
 package ddog.payment.application.exception;
 
 
-import ddog.payment.application.exception.common.CustomRuntimeException;
+import ddog.auth.exception.common.CustomRuntimeException;
 
 public class OrderException extends CustomRuntimeException {
     public OrderException(OrderExceptionType type, Object... args) {
