+++ daengle-payment-api/src/main/java/ddog/payment/application/exception/PaymentException.java
@@ -1,7 +1,7 @@
 package ddog.payment.application.exception;
 
 
-import ddog.payment.application.exception.common.CustomRuntimeException;
+import ddog.auth.exception.common.CustomRuntimeException;
 
 public class PaymentException extends CustomRuntimeException {
     public PaymentException(PaymentExceptionType type, Object... args) {
