--- daengle-payment-api/src/main/java/ddog/payment/application/exception/common/BadRequestException.java
@@ -1,9 +1,0 @@
-package ddog.payment.application.exception.common;
-
-import org.springframework.http.HttpStatus;
-
-public class BadRequestException extends CustomRuntimeException {
-    public BadRequestException(String message) {
-        super(message, HttpStatus.BAD_REQUEST, null);
-    }
-}
