--- daengle-payment-api/src/main/java/ddog/payment/application/exception/common/IllegalParameterException.java
@@ -1,7 +1,0 @@
-package ddog.payment.application.exception.common;
-
-public class IllegalParameterException extends RuntimeException {
-    public IllegalParameterException(String message, Object... args) {
-        super();
-    }
-}
