+++ daengle-user-api/src/main/java/ddog/user/application/exception/UserException.java
@@ -1,6 +1,6 @@
 package ddog.user.application.exception;
 
-import ddog.user.application.exception.common.CustomRuntimeException;
+import ddog.auth.exception.common.CustomRuntimeException;
 
 public class UserException extends CustomRuntimeException {
     public UserException(UserExceptionType type, Object... args) {
