--- daengle-vet-api/src/main/java/ddog/vet/application/exception/common/BadRequestException.java
@@ -1,9 +1,0 @@
-package ddog.vet.application.exception.common;
-
-import org.springframework.http.HttpStatus;
-
-public class BadRequestException extends CustomRuntimeException {
-    public BadRequestException(String message) {
-        super(message, HttpStatus.BAD_REQUEST, null);
-    }
-}
