--- daengle-vet-api/src/main/java/ddog/vet/application/exception/common/IllegalParameterException.java
@@ -1,7 +1,0 @@
-package ddog.vet.application.exception.common;
-
-public class IllegalParameterException extends RuntimeException {
-    public IllegalParameterException(String message, Object... args) {
-        super();
-    }
-}
